refactor(core): mode audio kernel commands to commands module
